+++ database/migrations/2022_02_09_134916_add_fiscal_code_to_customer.php
@@ -14,7 +14,7 @@ class AddFiscalCodeToCustomer extends Migration
     public function up()
     {
         Schema::table('customers', function (Blueprint $table) {
-            $table->string('fiscal_code', 16)->nullable(false);
+            $table->string('fiscal_code', 16)->nullable(true);
         });
     }